SQL и SQLite

+++ app/src/main/java/ru/netology/nmedia/dto/Post.kt
@@ -1,13 +1,13 @@
 package ru.netology.nmedia.dto
 
-data class Post (
+data class Post(
     val id: Long,
     val author: String,
     val content: String,
     val published: String,
-    val favorites: Long = 0,
-    var shares: Long = 0,
-    val views: Long = 0,
+    val favorites: Int = 0,
+    var shares: Int = 0,
+    val views: Int = 0,
     val favoritesByMe: Boolean,
     val sharesByMe: Boolean,
     val video: String?
